Controllers and routes as separate layers
